Add error handler to query operations
In previous versions the generated queries would crash the svelte application
on any error. This commit adds an error callback to the query.subscribe
invocation that assigns the apollo error to the `error` property on the store.
This is consistent with the behavior of the apollo client when used via
`client.query` or the `useQuery` react hooks.

When an error occurs, the `data` property is reset to the initial empty object.
This is problematic since it is inconsistent with the generated query return
type (this also applies to the initial store value supplied to the first call
to `readable`). An ideal solution is to set the initial/error values to `null`
and update the generated type to `data: T | null`, but this will likely cause
large numbers of (valid) type errors in existing applications.